pyback/requirements.txt
wget
Cython
numpy
sqlalchemy
scikit_image
pandas
pycocotools
matplotlib
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
mrcnn
colorama
keras
tensorflow
IPython 
gitPython
tqdm 
imutils
imgaug 
https://github.com/OlafenwaMoses/ImageAI/releases/download/2.0.3/imageai-2.0.3-py3-none-any.whl
